Backend/data context (#3)
* Checkpoint datacontext progress

* Add reminder TODO for internal API consideration

* Checkpoint: Initial CRUD support

We may want to expand the interface in future to enable easier batching, filtering, querying, etc. For now this will do though

* fixup: Match frontend_data models

* Fixup data upon import

This should probably be changed a bit when we have all our assets in
order as well as our finalised data json. For now, this will ensure
that we at least can point to images which will exist at runtime

* Cleanup: remove bookmark comments for IDE anchoring

* Doc: Document interface. Also define the nullability context of <get>

* Auto generate IDs conforming to IETF RFC4122 UUIDv7 spec

---------

Co-authored-by: sxha092024 <183362499+sxha092024@users.noreply.github.com>