Change ID to Guid
If loading data whose ID is not a Guid, but an otherwise previously valid record, perform a migration and re-assign ID.

Additionally, pull in an additional dependency on a background task service library to implement a background task which syncs/flushes data to disk at set periods. Yet another safeguard that valid data is stored.

As a side effect of the migration this commit *will* appear much large than it is, as 20 image files have been re-named as well.