Backend/api (#10)
* Change ID to Guid

If loading data whose ID is not a Guid, but an otherwise previously valid record, perform a migration and re-assign ID.

Additionally, pull in an additional dependency on a background task service library to implement a background task which syncs/flushes data to disk at set periods. Yet another safeguard that valid data is stored.

As a side effect of the migration this commit *will* appear much large than it is, as 20 image files have been re-named as well.

* Replace `.First` with `.Single` in LINQ

Ensure, and *critically* warn if the transactional logic has somehow created a situation with non-unique data in memory.

* Additional logging information on CRUD endpoints

* Pre-emptively handle model changes from commit 60ebd65

* Perform data model migration based on commit 60ebd65

* Update .gitignore

* Asset migration

---------

Co-authored-by: sxha092024 <183362499+sxha092024@users.noreply.github.com>